TCG/requirements.txt
Flask~=3.1.0
mysql-connector-python~=8.3.0
fastapi[all]~=0.115.6
uvicorn~=0.32.1
pyjwt~=2.10.1
openpyxl~=3.1.5
python-docx
PyPDF2~=3.0.1
pydantic~=2.10.3
openai~=1.77.0
requests
pandas~=2.2.3
tabulate
tcms-api
extract_msg
json-repair
tiktoken
aiohttp
sentence-transformers
